feat: Add Sign In Modal for FavoriteToggleButton
The code changes in this commit add a Sign In Modal to the FavoriteToggleButton component in the components/card/FavoriteToggleButton.tsx file. If there is no user logged in, the Sign In Modal will be displayed instead of the button. This improves the user experience by prompting users to sign in before interacting with the favorite toggle button.

Co-authored-by: [Author Name]