feat: Add fetchFavoriteId function for retrieving favorite ID
The code changes in this commit add a new function, `fetchFavoriteId`, to the `actions.ts` file in the `utils` directory. This function retrieves the ID of a favorite property for a given user and property ID. It queries the database to find the matching favorite entry and returns the ID if found, or null if not found. This function will be useful for retrieving the favorite ID when interacting with the favorite toggle button.