feat: Add lazy loading for images and Superhost information to UserInfo component
Refactor the UserInfo component in `components/properties/UserInfo.tsx` to include lazy loading for the profile image using the next/image package. This improves the page load time, especially for users with large profile images. Additionally, add the Superhost information to the component, displaying the number of years the user has been hosting. This enhancement provides users with more information about the host and improves the overall user experience of the application.